--- frontend/src/app/projects_listing/projects_enterprises/[project_id]/create/page.tsx
@@ -1,13 +1,0 @@
-import PEnterpriseCreateWrapper from '@ors/components/manage/Blocks/ProjectsListing/ProjectsEnterprises/create/PEnterpriseCreateWrapper'
-import PageWrapper from '@ors/components/theme/PageWrapper/PageWrapper'
-import usePageTitle from '@ors/hooks/usePageTitle'
-
-export default function PEnterprisesCreate() {
-  usePageTitle('Project enterprise create')
-
-  return (
-    <PageWrapper>
-      <PEnterpriseCreateWrapper />
-    </PageWrapper>
-  )
-}
