--- frontend/src/app/projects_listing/projects_enterprises/[project_id]/edit/[enterprise_id]/page.tsx
@@ -1,13 +1,0 @@
-import PEnterpriseEditWrapper from '@ors/components/manage/Blocks/ProjectsListing/ProjectsEnterprises/edit/PEnterpriseEditWrapper'
-import PageWrapper from '@ors/components/theme/PageWrapper/PageWrapper'
-import usePageTitle from '@ors/hooks/usePageTitle'
-
-export default function PEnterprisesEdit() {
-  usePageTitle('Project enterprise edit')
-
-  return (
-    <PageWrapper>
-      <PEnterpriseEditWrapper />
-    </PageWrapper>
-  )
-}
