--- frontend/src/app/projects_listing/projects_enterprises/[project_id]/view/[enterprise_id]/page.tsx
@@ -1,17 +1,0 @@
-import PEnterpriseViewWrapper from '@ors/components/manage/Blocks/ProjectsListing/ProjectsEnterprises/view/PEnterpriseViewWrapper'
-import PageWrapper from '@ors/components/theme/PageWrapper/PageWrapper'
-import usePageTitle from '@ors/hooks/usePageTitle'
-
-import { useParams } from 'wouter'
-
-export default function PEnterprise() {
-  usePageTitle('Project enterprise')
-
-  const { enterprise_id } = useParams<Record<string, string>>()
-
-  return (
-    <PageWrapper>
-      <PEnterpriseViewWrapper key={enterprise_id} />
-    </PageWrapper>
-  )
-}
