+++ frontend/src/components/manage/Blocks/ProjectsListing/Enterprises/create/EnterpriseCreateWrapper.tsx
@@ -4,17 +4,28 @@ import PermissionsContext from '@ors/contexts/PermissionsContext.tsx'
 import EnterpriseHeader from './EnterpriseHeader.tsx'
 import EnterpriseCreate from './EnterpriseCreate.tsx'
 import ProjectFormFooter from '../../ProjectFormFooter.tsx'
-import { initialOverviewFields } from '../../ProjectsEnterprises/constants.ts'
-import { EnterpriseOverview } from '../../interfaces.ts'
+import { EnterpriseData } from '../../interfaces.ts'
+import {
+  initialOverviewFields,
+  initialDetailsFields,
+  initialSubstanceFields,
+  initialFundingDetailsFields,
+  initialRemarksFields,
+} from '../../ProjectsEnterprises/constants.ts'
 
 import { Redirect } from 'wouter'
 
 const EnterpriseCreateWrapper = () => {
   const { canEditEnterprise } = useContext(PermissionsContext)
 
-  const [enterpriseData, setEnterpriseData] = useState<EnterpriseOverview>(
-    initialOverviewFields,
-  )
+  const [enterpriseData, setEnterpriseData] = useState<EnterpriseData>({
+    overview: initialOverviewFields,
+    details: initialDetailsFields,
+    substance_fields: initialSubstanceFields,
+    substance_details: [],
+    funding_details: initialFundingDetailsFields,
+    remarks: initialRemarksFields,
+  })
   const [enterpriseId, setEnterpriseId] = useState<number | null>(null)
 
   const [errors, setErrors] = useState<{ [key: string]: string[] }>({})
