+++ frontend/src/components/manage/Blocks/ProjectsListing/Enterprises/interfaces.ts
@@ -1,0 +1,4 @@
+export type StatusType = {
+  id: number
+  name: string
+}
