--- frontend/src/components/manage/Blocks/ProjectsListing/ProjectsEnterprises/tabs/PEnterpriseRemarksSection.tsx
@@ -1,43 +1,0 @@
-import { useContext } from 'react'
-
-import PermissionsContext from '@ors/contexts/PermissionsContext'
-import { EnterpriseTextAreaField } from '../FormHelperComponents'
-import { remarksFields } from '../constants'
-import {
-  PEnterpriseDataProps,
-  PEnterpriseData,
-  EnterpriseRemarks,
-} from '../../interfaces'
-
-import { map } from 'lodash'
-
-const PEnterpriseRemarksSection = ({
-  enterpriseData,
-  setEnterpriseData,
-  errors,
-}: PEnterpriseDataProps) => {
-  const { canEditProjectEnterprise } = useContext(PermissionsContext)
-  const isDisabled = !canEditProjectEnterprise
-
-  const sectionIdentifier = 'remarks'
-
-  return (
-    <div className="flex max-w-[41rem] flex-col gap-y-2">
-      {map(remarksFields, (field, index) => (
-        <EnterpriseTextAreaField<PEnterpriseData, EnterpriseRemarks>
-          key={index}
-          enterpriseData={enterpriseData.remarks}
-          {...{
-            setEnterpriseData,
-            sectionIdentifier,
-            field,
-            isDisabled,
-            errors,
-          }}
-        />
-      ))}
-    </div>
-  )
-}
-
-export default PEnterpriseRemarksSection
