--- frontend/src/components/manage/Blocks/ProjectsListing/hooks/useGetEnterpriseStatuses.ts
@@ -1,12 +1,0 @@
-import useApi from '@ors/hooks/useApi'
-
-export function useGetEnterpriseStatuses() {
-  const { data } = useApi({
-    options: {
-      withStoreCache: true,
-    },
-    path: '/api/enterprise-statuses/',
-  })
-
-  return data
-}
