--- frontend/src/components/manage/Blocks/ProjectsListing/hooks/useGetProjectEnterprise.ts
@@ -1,10 +1,0 @@
-import useApi from '@ors/hooks/useApi'
-
-export function useGetProjectEnterprise(enterprise_id: string) {
-  return useApi({
-    options: {
-      withStoreCache: false,
-    },
-    path: `api/project-enterprise/${enterprise_id}/`,
-  })
-}
